+++ app/routes/$character.tsx
@@ -3,8 +3,8 @@ import { json } from "@remix-run/node";
 import { prefs } from "./../utils/cookies";
 import { useLoaderData, useFetcher, useSubmit, useOutlet, useRevalidator, useNavigate } from "@remix-run/react";
 import type { MetaFunction } from "@remix-run/node";
-import type { Event } from "./$character.events";
-import { EventLog } from "./$character.events";
+import type { Event } from "./$character_.events";
+import { EventLog } from "./$character_.events";
 import type { Post } from "./$character_.posts";
 import { PostLog } from "./$character_.posts";
 import type { Message } from "./$character.chat.$thread";
